requirements.txt
flask
pydantic
pytest
requests
allure-pytest
